pyproject.toml
[tool.poetry]
name = "mt_mesonet_satellite"
version = "0.1.0"
description = ""
authors = ["Colin Brust <colin.brust@mso.umt.edu>"]
license = "MIT"
packages = [
    { include = "mt_mesonet_satellite" }
]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.27.1"
pandas = "^1.4.2"
numpy = "^1.22.4"
geopandas = "^0.10.2"
pyjanitor = "^0.23.1"
neo4j = "^4.4.4"
neo4j-driver = "^4.4.4"
python-dotenv = "^0.20.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
ipykernel = "^6.13.0"
matplotlib = "^3.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
